requirements.txt
selenium==4.32.0
openai==1.78.0
beautifulsoup4==4.13.4
python-dotenv==1.1.0
